# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "MovieFlix API",
    "description": "API para gerenciamento de filmes do MovieFlix",
    "version": "1.0.0"
  },
  "paths": {
    "/movies": {
      "get": {
        "tags": ["Movie"],
        "summary": "Listar todos os filmes",
        "responses": {
          "200": {
            "description": "Lista de filmes retornada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Movie"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Movie"],
        "summary": "Cadastrar um novo filme",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MovieInput"
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Filme cadastrado com sucesso" },
          "409": {
            "description": "Já existe um filme com esse título",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "500": {
            "description": "Falha ao cadastrar o filme",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          }
        }
      }
    },
    "/movies/{id}": {
      "put": {
        "tags": ["Movie"],
        "summary": "Atualizar um filme",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" },
            "description": "ID do filme a ser atualizado"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MovieInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Filme atualizado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "404": {
            "description": "Filme não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "500": {
            "description": "Falha ao atualizar o filme",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Movie"],
        "summary": "Remover um filme",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" },
            "description": "ID do filme a ser removido"
          }
        ],
        "responses": {
          "200": {
            "description": "Filme deletado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "404": {
            "description": "Filme não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "500": {
            "description": "Falha ao remover o filme",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          }
        }
      }
    },
    "/movies/{genreName}": {
      "get": {
        "tags": ["Movie"],
        "summary": "Filtrar filmes por gênero",
        "parameters": [
          {
            "name": "genreName",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "description": "Nome do gênero para filtrar os filmes"
          }
        ],
        "responses": {
          "200": {
            "description": "Filmes filtrados por gênero retornados com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Movie"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Falha ao filtrar filmes por gênero",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Genre": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "ID do gênero"
          },
          "name": {
            "type": "string",
            "description": "Nome do gênero"
          }
        }
      },
      "Language": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "ID do idioma"
          },
          "name": {
            "type": "string",
            "description": "Nome do idioma"
          }
        }
      },
      "Movie": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "ID do filme"
          },
          "title": {
            "type": "string",
            "description": "Título do filme"
          },
          "release_date": {
            "type": "string",
            "format": "date",
            "description": "Data de lançamento do filme"
          },
          "genre_id": {
            "type": "integer",
            "description": "ID do gênero do filme"
          },
          "language_id": {
            "type": "integer",
            "description": "ID do idioma do filme"
          },
          "oscar_count": {
            "type": "integer",
            "description": "Número de Oscars ganhos pelo filme"
          },
          "genres": {
            "$ref": "#/components/schemas/Genre"
          },
          "languages": {
            "$ref": "#/components/schemas/Language"
          }
        }
      },
      "MovieInput": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Título do filme"
          },
          "genre_id": {
            "type": "integer",
            "description": "ID do gênero do filme"
          },
          "language_id": {
            "type": "integer",
            "description": "ID do idioma do filme"
          },
          "oscar_count": {
            "type": "integer",
            "description": "Número de Oscars ganhos pelo filme"
          },
          "release_date": {
            "type": "string",
            "format": "date",
            "description": "Data de lançamento do filme"
          }
        },
        "required": ["title", "genre_id", "language_id", "oscar_count", "release_date"]
      },
      "MessageResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Mensagem de resposta"
          }
        }
      }
    }
  }
}